unified mock versions of SKProduct
