--- PurchasesTests/Mocks/MockDiscount.swift
@@ -1,6 +1,0 @@
-//
-// Created by Andrés Boedo on 6/3/20.
-// Copyright (c) 2020 Purchases. All rights reserved.
-//
-
-import Foundation
