+++ PurchasesTests/PurchasesTests.swift
@@ -270,7 +270,7 @@ class PurchasesTests: XCTestCase {
         let purchaserInfo = Purchases.PurchaserInfo()
         self.backend.postReceiptPurchaserInfo = purchaserInfo
 
-        let product = MockProduct(mockProductIdentifier: "product")
+        let product = MockSKProduct(mockProductIdentifier: "product")
         let payment = SKPayment(product: product)
 
         let transaction = MockTransaction()
@@ -300,7 +300,7 @@ class PurchasesTests: XCTestCase {
 
         let purchaserInfo2 = purchaserInfo1
 
-        let product = MockProduct(mockProductIdentifier: "product")
+        let product = MockSKProduct(mockProductIdentifier: "product")
         let payment = SKPayment(product: product)
 
         let transaction = MockTransaction()
@@ -341,7 +341,7 @@ class PurchasesTests: XCTestCase {
             ]
         ])
 
-        let product = MockProduct(mockProductIdentifier: "product")
+        let product = MockSKProduct(mockProductIdentifier: "product")
         let payment = SKPayment(product: product)
 
         let transaction = MockTransaction()
@@ -365,7 +365,7 @@ class PurchasesTests: XCTestCase {
 
     func testDelegateIsNotCalledIfBlockPassed() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -403,7 +403,7 @@ class PurchasesTests: XCTestCase {
 
     func testAddsPaymentToWrapper() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -414,7 +414,7 @@ class PurchasesTests: XCTestCase {
 
     func testTransitioningToPurchasing() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -430,7 +430,7 @@ class PurchasesTests: XCTestCase {
 
     func testTransitioningToPurchasedSendsToBackend() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -450,7 +450,7 @@ class PurchasesTests: XCTestCase {
 
     func testReceiptsSendsAsRestoreWhenAnon() {
         setupAnonPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -471,7 +471,7 @@ class PurchasesTests: XCTestCase {
     func testReceiptsSendsAsNotRestoreWhenAnonymousAndNotAllowingSharingAppStoreAccount() {
         setupAnonPurchases()
         self.purchases.allowSharingAppStoreAccount = false
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -492,7 +492,7 @@ class PurchasesTests: XCTestCase {
     func testReceiptsSendsAsRestoreWhenNotAnonymousAndAllowingSharingAppStoreAccount() {
         setupPurchases()
         self.purchases.allowSharingAppStoreAccount = true
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -512,7 +512,7 @@ class PurchasesTests: XCTestCase {
 
     func testFinishesTransactionsIfSentToBackendCorrectly() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -535,7 +535,7 @@ class PurchasesTests: XCTestCase {
     func testDoesntFinishTransactionsIfFinishingDisabled() {
         setupPurchases()
         self.purchases?.finishTransactions = false
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -608,7 +608,7 @@ class PurchasesTests: XCTestCase {
 
     func testFetchesProductInfoIfNotCached() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
 
         let transaction = MockTransaction()
         storeKitWrapper.payment = SKPayment(product: product);
@@ -636,7 +636,7 @@ class PurchasesTests: XCTestCase {
 
     func testAfterSendingDoesntFinishTransactionIfBackendError() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -654,7 +654,7 @@ class PurchasesTests: XCTestCase {
 
     func testAfterSendingFinishesFromBackendErrorIfAppropriate() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -673,7 +673,7 @@ class PurchasesTests: XCTestCase {
 
     func testNotifiesIfTransactionFailsFromBackend() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -692,7 +692,7 @@ class PurchasesTests: XCTestCase {
 
     func testNotifiesIfTransactionFailsFromStoreKit() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         var receivedError: Error?
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
             receivedError = error
@@ -714,7 +714,7 @@ class PurchasesTests: XCTestCase {
 
     func testCallsDelegateAfterBackendResponse() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         
         var purchaserInfo: Purchases.PurchaserInfo?
         var receivedError: Error?
@@ -742,7 +742,7 @@ class PurchasesTests: XCTestCase {
 
     func testCompletionBlockOnlyCalledOnce() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
 
         var callCount = 0
 
@@ -765,8 +765,8 @@ class PurchasesTests: XCTestCase {
 
     func testCompletionBlockNotCalledForDifferentProducts() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
-        let otherProduct = MockProduct(mockProductIdentifier: "com.product.id2")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
+        let otherProduct = MockSKProduct(mockProductIdentifier: "com.product.id2")
 
         var callCount = 0
 
@@ -788,7 +788,7 @@ class PurchasesTests: XCTestCase {
 
     func testCallingPurchaseWhileSameProductPendingIssuesError() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
 
         // First one "works"
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
@@ -937,7 +937,7 @@ class PurchasesTests: XCTestCase {
 
     func testCallsShouldAddPromoPaymentDelegateMethod() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "mock_product")
+        let product = MockSKProduct(mockProductIdentifier: "mock_product")
         let payment = SKPayment.init()
 
         storeKitWrapper.delegate?.storeKitWrapper(storeKitWrapper, shouldAddStore: payment, for: product)
@@ -947,7 +947,7 @@ class PurchasesTests: XCTestCase {
 
     func testShouldAddPromoPaymentDelegateMethodReturnsFalse() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "mock_product")
+        let product = MockSKProduct(mockProductIdentifier: "mock_product")
         let payment = SKPayment.init()
 
         let result = storeKitWrapper.delegate?.storeKitWrapper(storeKitWrapper, shouldAddStore: payment, for: product)
@@ -957,7 +957,7 @@ class PurchasesTests: XCTestCase {
 
     func testShouldCacheProductsFromPromoPaymentDelegateMethod() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "mock_product")
+        let product = MockSKProduct(mockProductIdentifier: "mock_product")
         let payment = SKPayment.init(product: product)
 
         storeKitWrapper.delegate?.storeKitWrapper(storeKitWrapper, shouldAddStore: payment, for: product)
@@ -978,7 +978,7 @@ class PurchasesTests: XCTestCase {
 
     func testDeferBlockMakesPayment() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "mock_product")
+        let product = MockSKProduct(mockProductIdentifier: "mock_product")
         let payment = SKPayment.init(product: product)
 
         storeKitWrapper.delegate?.storeKitWrapper(storeKitWrapper, shouldAddStore: payment, for: product)
@@ -1070,7 +1070,7 @@ class PurchasesTests: XCTestCase {
                 "other_purchases": [:]
             ]]);
 
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -1557,7 +1557,7 @@ class PurchasesTests: XCTestCase {
 
     func testUserCancelledFalseIfPurchaseSuccessful() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         var receivedUserCancelled: Bool?
 
         // Second one issues an error
@@ -1575,7 +1575,7 @@ class PurchasesTests: XCTestCase {
 
     func testUserCancelledTrueIfPurchaseCancelled() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         var receivedUserCancelled: Bool?
         var receivedError: NSError?
         var receivedUnderlyingError: NSError?
@@ -1604,7 +1604,7 @@ class PurchasesTests: XCTestCase {
         setupPurchases()
         self.receiptFetcher.shouldReturnReceipt = false
 
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         var receivedUserCancelled: Bool?
         var receivedError: NSError?
 
@@ -1625,7 +1625,7 @@ class PurchasesTests: XCTestCase {
 
     func testDeferBlockCallsCompletionBlockAfterPurchaseCompletes() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "mock_product")
+        let product = MockSKProduct(mockProductIdentifier: "mock_product")
         let payment = SKPayment.init(product: product)
 
         storeKitWrapper.delegate?.storeKitWrapper(storeKitWrapper, shouldAddStore: payment, for: product)
@@ -1651,7 +1651,7 @@ class PurchasesTests: XCTestCase {
     func testAddsDiscountToWrapper() {
         if #available(iOS 12.2, *) {
             setupPurchases()
-            let product = MockProduct(mockProductIdentifier: "com.product.id1")
+            let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
             let discount = SKPaymentDiscount.init(identifier: "discount", keyIdentifier: "TIKAMASALA1", nonce: UUID(), signature: "Base64 encoded signature", timestamp: 123413232131)
 
             self.purchases?.purchaseProduct(product, discount: discount) { (tx, info, error, userCancelled) in
@@ -1667,7 +1667,7 @@ class PurchasesTests: XCTestCase {
     func testPaymentDiscountForProductDiscountCreatesDiscount() {
         if #available(iOS 12.2, *) {
             setupPurchases()
-            let product = MockProduct(mockProductIdentifier: "com.product.id1")
+            let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
 
             let discountIdentifier = "id"
             let signature = "firma"
@@ -1779,7 +1779,7 @@ class PurchasesTests: XCTestCase {
 
     func testObserverModeSetToFalseSetFinishTransactions() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -1801,7 +1801,7 @@ class PurchasesTests: XCTestCase {
 
     func testDoesntFinishTransactionsIfObserverModeIsSet() {
         setupPurchasesObserverModeOn()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -1823,7 +1823,7 @@ class PurchasesTests: XCTestCase {
 
     func testRestoredPurchasesArePosted() {
         setupPurchasesObserverModeOn()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -1907,7 +1907,7 @@ class PurchasesTests: XCTestCase {
         let purchaserInfo = Purchases.PurchaserInfo()
         self.backend.postReceiptPurchaserInfo = purchaserInfo
 
-        let product = MockProduct(mockProductIdentifier: "product")
+        let product = MockSKProduct(mockProductIdentifier: "product")
         let payment = SKPayment(product: product)
 
         let transaction = MockTransaction()
@@ -1926,7 +1926,7 @@ class PurchasesTests: XCTestCase {
     
     func testReceiptsSendsObserverModeWhenObserverMode() {
         setupPurchasesObserverModeOn()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
@@ -1946,7 +1946,7 @@ class PurchasesTests: XCTestCase {
     
     func testReceiptsSendsObserverModeOffWhenObserverModeOff() {
         setupPurchases()
-        let product = MockProduct(mockProductIdentifier: "com.product.id1")
+        let product = MockSKProduct(mockProductIdentifier: "com.product.id1")
         self.purchases?.purchaseProduct(product) { (tx, info, error, userCancelled) in
 
         }
