+++ PurchasesTests/StoreKitRequestFetcherTests.swift
@@ -13,66 +13,14 @@ import StoreKit
 
 import Purchases
 
-@available(iOS 11.2, *)
-class MockDiscount: SKProductDiscount {
-    override var paymentMode: SKProductDiscount.PaymentMode {
-        return SKProductDiscount.PaymentMode.payAsYouGo
-    }
-
-    override var price: NSDecimalNumber {
-        return 1.99 as NSDecimalNumber
-    }
-}
-
-class MockProduct: SKProduct {
-
-    var mockProductIdentifier: String
-
-    init(mockProductIdentifier: String) {
-        self.mockProductIdentifier = mockProductIdentifier
-        super.init()
-    }
-
-    override var productIdentifier: String {
-        return self.mockProductIdentifier
-    }
-    
-    var mockSubscriptionGroupIdentifier: String?
-    override var subscriptionGroupIdentifier: String? {
-        return self.mockSubscriptionGroupIdentifier;
-    }
-
-    override var priceLocale: Locale {
-        return Locale.current
-    }
-
-    override var price: NSDecimalNumber {
-        return 2.99 as NSDecimalNumber
-    }
-
-    @available(iOS 11.2, *)
-    override var introductoryPrice: SKProductDiscount? {
-        return MockDiscount()
-    }
-    
-    var mockDiscountIdentifier: String?
-    @available(iOS 12.2, *)
-    override var discounts: [SKProductDiscount] {
-        if (mockDiscountIdentifier != nil) {
-            return [MockProductDiscount(mockIdentifier: mockDiscountIdentifier!)];
-        } else {
-            return []
-        }
-    }
-}
 
 class StoreKitRequestFetcher: XCTestCase {
 
     class MockProductResponse: SKProductsResponse {
-        var mockProducts: [MockProduct]
+        var mockProducts: [MockSKProduct]
         init(productIdentifiers: Set<String>) {
             self.mockProducts = productIdentifiers.map { identifier in
-                return MockProduct(mockProductIdentifier: identifier)
+                return MockSKProduct(mockProductIdentifier: identifier)
             }
             super.init()
         }
